Fix xlog reporting bug affecting nh4/fourk/etc.
game["duration_str"] was being set to game["realtime"] if "wallclock" in
game and "realtime" not in game, when it should be game["wallclock"]

thanks to @entrez for spotting this

seems the incorrect handling of that conditional had a side-effect of
breaking with some xlogs for games where "wallclock" exists but
"realtime" doesn't. will test to see if this fixes the problem.

Signed-off-by: Joanna Doyle <jjadoyle@gmail.com>