Hardcode minimum turncounts by playername within botconf.py
Some players can be spammy with their games, purposely dying in a short
amount of time but not actually meeting the threshold for startscumming.
This commit allows admins with edit access to botconf.py to hardcode
player names and set a minimum turn count for that player that can't be
overridden by the !setmintc command, not even by a channel admin. A
message is sent privately to the player/admin that tries to use
!setmintc for a player name that is set within the botconf.py
!configuration file.